Big refactor to allow porting to linux. Now most things should work properly on linux, may have messed up some windows stuff and have to fix them.
